refactor(tauri): extract the pending-update marker into its own module
First cluster of the lib.rs split (Unit 11a). A pure move: every
non-comment line is verbatim, verified line by line against the removed
text, and the 18 tests move with it as a descendant module so they keep
private-field access to PendingUpdateMarker without widening anything.

Only what lib.rs still calls is pub. check_for_update_now deliberately did
NOT move: its body is one call to check_for_update, a private crate-root
function five thousand lines away that four unrelated sites also use, so
moving the command would have forced that function public — a visibility
widening on an item outside the cluster, which is what the review question
asks about.

cargo fmt is NOT run. It is not gated in CI and this repo is not
rustfmt-clean: running it reformatted sixteen untouched files and buried
this move under ~500 lines of unrelated churn.

Call site, thread and ordering are unchanged — evaluate_pending_update_
marker still runs synchronously in setup(), before the async spawn, for
the reason its own doc comment gives.
